Refactor to abstract `CaptchaListener` class
