+++ src/Listeners/ValidateFormSubmission.php
@@ -2,35 +2,17 @@
 
 namespace AryehRaber\Captcha\Listeners;
 
-use AryehRaber\Captcha\Captcha;
 use Statamic\Events\FormSubmitted;
 use Statamic\Forms\Submission;
 
-class ValidateFormSubmission
+class ValidateFormSubmission extends CaptchaListener
 {
-    protected $captcha;
-
-    public function __construct(Captcha $captcha)
-    {
-        $this->captcha = $captcha;
-    }
-
-    public function handle(FormSubmitted $event)
+    /** @param FormSubmitted $event */
+    protected function shouldVerify($event): bool
     {
         /** @var Submission */
         $submission = $event->submission;
 
-        if (! $this->shouldVerify($submission)) {
-            return null;
-        }
-
-        $this->captcha->verify()->throwIfInvalid();
-
-        return null;
-    }
-
-    protected function shouldVerify(Submission $submission)
-    {
         $shouldVerify = config('captcha.forms') === 'all'
             || in_array($submission->form()->handle(), config('captcha.forms', []));
