+++ src/Listeners/ValidateUserLogin.php
@@ -2,32 +2,12 @@
 
 namespace AryehRaber\Captcha\Listeners;
 
-use AryehRaber\Captcha\Captcha;
 use Illuminate\Auth\Events\Login;
 
-class ValidateUserLogin
+class ValidateUserLogin extends CaptchaListener
 {
-    protected $captcha;
-
-    public function __construct(Captcha $captcha)
-    {
-        $this->captcha = $captcha;
-    }
-
-    public function handle(Login $event)
-    {
-        $user = $event->user;
-
-        if (! $this->shouldVerify()) {
-            return null;
-        }
-
-        $this->captcha->verify()->throwIfInvalid();
-
-        return null;
-    }
-
-    protected function shouldVerify()
+    /** @param Login $event */
+    protected function shouldVerify($event): bool
     {
         return config('captcha.user_login', false);
     }
